Restructure folder as package
