+++ perception/sandbox/eyehandcal/readme.md
@@ -1,5 +1,12 @@
 # Understanding Eye Hand Calibraton 
 
+## Setup
+
+```bash
+conda env create -f environment.yml
+conda activate eyehandcal  # or direnv allow
+```
+
 
 ## Assumption
 * Using a realsense camera
